+++ src/find_max_min_med.c
@@ -6,23 +6,23 @@
 /*   By: mproveme <mproveme@student.21-school.ru    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/25 18:15:35 by mproveme          #+#    #+#             */
-/*   Updated: 2022/03/26 18:09:39 by mproveme         ###   ########.fr       */
+/*   Updated: 2022/03/27 14:59:05 by mproveme         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header/push_swap.h"
 
-int	get_max(long *arr, int len)
+int	get_max(int *arr, int len)
 {
 	return (arr[len - 1]);
 }
 
-int	get_min(long *arr)
+int	get_min(int *arr)
 {
 	return (arr[0]);
 }
 
-int	get_med(long *arr, int len)
+int	get_med(int *arr, int len)
 {
 	int	x;
